Fix MP3 playback
The define line was taken from the Chinese GMU RetroFW port (thanks @guangmingwan). That change on it's own proved to be not enough to fix the problem though.

The real fix was to change the 'mpg123_set_filesize' and 'mpg123_length' functions to 'mpg123_set_filesize_64' and 'mpg123_length_64' (in combination with the previous fix), the latter ones are defined as symbols in the mpg123 library but the former ones are not, so my best guess is that GMU was written for a previous version of the library (needs further investigation). The only side effect of this is GCC throwing some "implicit declaration of function" warnings.